+++ admin/wysiwyg/htmlarea.js
@@ -462,8 +462,7 @@ var node=obj[n];if(typeof node == 'object'){newObj[n]=HTMLArea.cloneObject(node)
 else { newObj[n] = node;}} return newObj;};HTMLArea.checkSupportedBrowser = function(){
 if(HTMLArea.is_gecko){if(navigator.productSub < 20021201){alert("You need at least Mozilla-1.3 Alpha.\n"+
 "Sorry, your Gecko is not supported.");return false;}
-if(navigator.productSub < 20030210){alert("Mozilla < 1.3 Beta is not supported!\n"+
-"I'll try, though, but it might not work.");}}
+}
 return HTMLArea.is_gecko || HTMLArea.is_ie;};HTMLArea.prototype._getSelection = function(){
 if(HTMLArea.is_ie){return this._doc.selection;}else{return this._iframe.contentWindow.getSelection();}};
 HTMLArea.prototype._createRange = function(sel){if (HTMLArea.is_ie){